requirements.txt
aiobotocore==2.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aioitertools==0.12.0
aiosignal==1.3.2
alabaster==1.0.0
anndata==0.11.3
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
appnope==0.1.4
array_api_compat==1.11.1
arrow==1.3.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.3
bibtexparser==1.4.3
botocore==1.37.3
cellxgene-census==1.17.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
docutils==0.21.2
duckduckgo_search==7.5.2
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
fake-useragent==2.1.0
filelock==3.18.0
fonttools==4.56.0
free-proxy==1.1.3
frozenlist==1.5.0
fsspec==2025.3.2
fuzzywuzzy==0.18.0
google-search-results==2.4.2
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.10
igraph==0.11.8
imagesize==1.4.1
ipykernel==6.29.5
ipython==8.34.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langchain==0.3.21
langchain-anthropic==0.3.10
langchain-community==0.3.20
langchain-core==0.3.45
langchain-openai==0.3.9
langchain-text-splitters==0.3.7
langsmith==0.3.17
legacy-api-wrap==1.4.1
leidenalg==0.10.2
llvmlite==0.44.0
lxml==5.3.1
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.2.0
mypy-extensions==1.0.0
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai==1.66.5
openpyxl==3.1.5
orjson==3.10.15
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
primp==0.14.0
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyarrow-hotfix==0.7
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
Pympler==1.1
pynndescent==0.5.13
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
RapidFuzz==3.12.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
s3fs==2025.3.2
safetensors==0.5.3
scanpy==1.11.0
scholarly==1.7.11
scikit-learn==1.5.2
scipy==1.15.2
seaborn==0.13.2
selenium==4.29.0
sentence-transformers==3.4.1
session-info2==0.1.2
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
somacore==1.0.28
sortedcontainers==2.4.0
soupsieve==2.6
Sphinx==8.1.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.39
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.13.1
tenacity==9.0.0
texttable==1.7.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tiledbsoma==1.16.2
tokenizers==0.21.1
tomli==2.2.1
torch==2.6.0
torchdata==0.11.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.0
trio==0.29.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
umap-learn==0.5.7
urllib3==2.3.0
wcwidth==0.2.13
websocket-client==1.8.0
wikipedia==1.4.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zstandard==0.23.0
